add authentication and authorization to the api

+++ controllers/api/v1/posts_api.js
@@ -24,13 +24,21 @@ module.exports.destroy = async function(req, res){
     try{
         let post = await Post.findById(req.params.id);
 
-        post.remove();
-    
-        await Comment.deleteMany({post: req.params.id});
+        if(post.user == req.user.id){
+
+            post.remove();    
+            await Comment.deleteMany({post: req.params.id});
+            return res.json(200, {
+                message: "Post and associated comments deleted successfully"
+            });
+
+        }else{
+
+            return res.json(401, {
+                message: "You cannot delete this post"
+            });
             
-        return res.json(200, {
-            message: "Post and associated comments deleted successfully"
-        });
+        }
 
     }catch(err){
         return res.json(500, {
